Rewrote the Duration Class.
Duration() is now a subclass of datetime.timdelta. All supported
(private) methods of parsing now return a dictionary of keyword
arguments compatible with the datetime.timedelta constructor. Added the
ability to parse generic objects, is possible. The class is also much
faster at parsing strings. Also added the timeFunction method to the
Timer class, which can benchmark a pre-defined funciton.